create Requests for News, Categories

+++ app/Http/Controllers/CategoriesController.php
@@ -15,7 +15,11 @@ class CategoriesController extends Controller
 
     public function categoriesById($id)
     {
-        return response()->json(Category::find($id)->load('news'), 200);
+        $category = Category::find($id);
+        if (is_null($category)){
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
+        return response()->json($category->load('news'), 200);
     }
 
     public function categoriesStore(Request $request)
@@ -24,14 +28,22 @@ class CategoriesController extends Controller
         return response()->json($category, 201);
     }
 
-    public function categoriesUpdate(Request $request, Category $category)
+    public function categoriesUpdate(Request $request, $id)
     {
+        $category = Category::find($id);
+        if (is_null($category)){
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $category->update($request->all());
         return response()->json($category, 200);
     }
 
-    public function categoriesDestroy(Category $category)
+    public function categoriesDestroy($id)
     {
+        $category = Category::find($id);
+        if (is_null($category)){
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $category->delete();
         return response()->json('', 204); 
     }
